+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog
 1.0.0 (unreleased)
 ------------------
 
+- Made compliant with Plone 4, Plone 5, Plone 6
+  [boulch, laulaz, sgeulette]
 - Require `pathlib2` in `setup.py`, backport `pathlib` for `py2.7`.
   [gbastien]
 - Added `security.setup_app` to be used in run scripts.
@@ -11,6 +13,8 @@ Changelog
 - Added `setup.load_xml_tool_only_from_package` to load only main tool xml file.
   [sgeulette]
 - Added `setup.test_remove_gs_step` to remove a generic setup step.
+  [sgeulette]
+- Added `imio.helpers.YesNoForFacetedVocabulary`
   [sgeulette]
 
 0.80 (2023-12-11)