+++ CHANGES.rst
@@ -12,6 +12,9 @@ Changelog
   [sgeulette]
 - Added `setup.test_remove_gs_step` to remove a generic setup step.
   [sgeulette]
+- Fixed `content.base_getattr` that was not returning the `default` if attribute
+  not existing.
+  [gbastien]
 
 0.80 (2023-12-11)
 -----------------